+++ src/calendar.h
@@ -84,7 +84,7 @@ class calendar
         /** Returns true if it's currently after sunset + TWILIGHT_SECONDS or before sunrise. */
         bool is_night() const;
         /** Returns the current sunlight or moonlight level through the preceding functions. */
-        int sunlight() const;
+        float sunlight() const;
 
         /** Basic accessors */
         int seconds() const
